+++ daemon/utmpd/utmpserver3.c
@@ -45,7 +45,7 @@ int action_frequently(int uid)
 	flood_base_hour=hour;
     }
 
-    if(abs(flooding[uid].lastlogin-(unsigned short)now)<=3 ||
+    if(abs((int)flooding[uid].lastlogin - (int)(unsigned short)now) <= 3 ||
 	    flooding[uid].minute_count>10 ||
 	    flooding[uid].hour_count>60) {
 	count_flooding++;
